This version is broken - use the alpha1 tag for the time being if you want something that works.
This version has the start of an attempt to add FAT32 alongside FAT16
support

UART stuff is fiddled about a bit to build smaller

Code is modified a little for size

Debug directories are added to hold the AS6 auto-generated Makefiles so it
can be built at the command line.

\README.TXT is added if for no other reason than to give a link back to
the project website.