Revision 13
Wouldn't you know it. The whole problem was caused by me making the main()
in .init3 "naked". I have a lot of stack frame variables but almost all of
them are used so locally that the compiler simply cached them into
registers EXCEPT for fat_offset which, because it's set long before it's
used, was a stack frame variable. Only when you have said you don't want a
stack frame things don't really add up. What a bugger of a bug it was to
find!

Oh and I moved the CRC stuff so it can be done twice. Not only before the
app code is launched, but right after the sectors have been read and
programmed. If this check fails then the EEPROM var is left at 0xFFFF
rather than marking the app as valid with a version number. This not only
prevents rubbish being used when some berk forgets to create a stack frame
but also it makes sure that the prepared AVRApnnn on disk really does have
an embedded checksum.