Revision 16
Fixed some problems in the FAT following (once it reached a cluster
boundary it was then doing the chain read for every sector after)

Added tentative support for FAT32 following but tricky to test it on a 32K
micro when FAT32 (4GB) clusters are 32K so trying to span one is tricky -
so it's theoretical for the time being.

Remove call to crc_app_ok() when CRC_FLASH not enabled

Make more stuff configurable so there are some combinations that can fit
into the 2K of smaller devices.