Revision 21
A number of changes after feedback from user sswift (thank you) who tried
to use the code and raised questions in several areas.

I have added copies of srec_cat.exe and avr-source.exe to the project so
its not reliant on sourcing these elsewhere.

I added a lot of comments to explain in more detail various parts of the
code.

I changed the LED flashing stuff to use a specific bit on a specific port
that can be #defined at the top of the file.

I had a debug routine to flood RAM with a known value at the start - I've
made this optional (and it only hard-codes RAMSTART if not defined now).