Revision 22
Main change is to add ROOT_MULTISEC support so that more than one root
sector will be searched for the AVRAP* file. When enabled it scans across
a whole cluster full of sectors (SecPerClus)

main.c is re-arranged to put support functions at the end and remove a lot
of the #define configuration and constants to other .h files

configure.h is added to the project to collect together all build
configuration in one easily editable package

The SPI pins used are now defined in configure.h and changed to be PB4..7
on mega16/32 which are the same pins as the hardware SPI (so the app code
can take over using the SD with real SPI when it starts).

All files are de-tabify'd to replace tabs with 4 spaces so
layout/alignment should be the same in any editor

A small change in uart.* that recognises that UART_puts() only make sense
from PSTR() so the existing UART_putsP is changed slightly.